backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.0.1
pymilvus==2.3.8
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
sentence-transformers==3.4.1
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
validators==0.34.0
openai>=1.68.0